# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0", 
        "title": "My User Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/stations": {
            "get": {
                "tags": ["Stations Route"],
                "description": "Get all stations in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User token to validate",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                ]
            }
        },
        "/stations/register": {
            "post": {
                "tags": ["Stations Route"],
                "description": "Allow users to register as a new stations \n name: [type: string ],  \n identificator: [type: string ],  \n longitude: [type: string ],  \n latitude: [type: string ],  \n altitude: [type: string ], \n userId: [type: string ], \n number: [type: integer ]",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User token to validate",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                ]
            }
        },
...
        "/stations/{id}": {
            "delete": {
                "tags": ["Stations Route"],
                "description": "Allow users to delete as a stations ",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User token to validate",
                        "schema": {
                            "$ref": "#/"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id to delete",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                ]
            }
        },
        "/stations/infos/{identificator}": {
            "post": {
                "tags": ["Stations Route"],
                "description": "Allow users to Update stations values \n values: \n [ \n type: array,  \n temperature: [type: string ],  \n windSpeed: [type: string ],  \n humidity: [type: string ],  \n dateDeMesure: [type: string ] \n ]",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User token to validate",
                        "schema": {
                            "$ref": "#/"
                        }
                    },
                    {
                        "name": "identificator",
                        "in": "path",
                        "description": "stations parameters",
                        "schema": {
                            "$ref": "#/"
                        }
                    },
                    {
                        "name": "params",
                        "in": "body",
                        "description": "Stations params",
                        "schema": {
                            "$ref": "#/"
                        }
                    }
                ]
            }
        },
...
    },
...
}